% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytuesday2026Jan.R
\docType{data}
\name{tidytuesday2026Jan}
\alias{tidytuesday2026Jan}
\title{tidytuesday data in 2026 Jan}
\format{
A remote object that returns a data frame.
}
\usage{
tidytuesday2026Jan
}
\description{
Data from the \href{https://github.com/rfordatascience/tidytuesday}{TidyTuesday project}
in 2026 up to 2026-01-20.
}
\keyword{datasets}
